[master] uint16 -> uint32 fix

+++ session.go
@@ -279,7 +279,7 @@ func (s *Session) sendLoop() {
 		case request := <-s.writes:
 			buf[0] = request.frame.ver
 			buf[1] = request.frame.cmd
-			binary.LittleEndian.PutUint32(buf[2:], uint16(len(request.frame.data)))
+			binary.LittleEndian.PutUint32(buf[2:], uint32(len(request.frame.data)))
 			binary.LittleEndian.PutUint32(buf[6:], request.frame.sid)
 			copy(buf[headerSize:], request.frame.data)
 			n, err := s.conn.Write(buf[:headerSize+len(request.frame.data)])
